server/pyproject.toml
[project]
name = "fanyi-server"
version = "0.1.0"
description = "Local translation gateway for fanyi-ext extension"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.30",
    "httpx>=0.27",
    "pyyaml>=6.0",
    "orjson>=3.10",
    "python-multipart>=0.0.26",
]

[tool.uv]
dev-dependencies = ["pytest>=8"]
